Add MIT License to the project
In this commit, an MIT license file has been added to the root of the project. The addition of this license provides users with the legal framework pertaining to their rights to use, copy, modify, merge, publish, distribute this software, and to permit others to do so under specified conditions.